refactor(core): specify explicit file paths in imports
